refactor(mobile): register walletKitListeners and strip them from provider
Wire walletKitListeners into the store's listener array and remove the
propose-response and safe-switch useEffects from WalletKitProvider, which
now owns only SDK setup + subscriptions. Listeners are registered globally
but stay inert when the feature is off (nothing populates the outstanding/
pending state). Migrate the provider listener test to a listeners-owned
test that registers the middleware directly, plus new coverage for the
cancelRequested, markReviewAbandoned, and rejectPending paths.
